requirements.txt
faiss-gpu==1.7.2
matplotlib==3.6.2
numpy==1.23.5
opencv_python_headless==4.8.1.78
scikit-learn==1.2.2
scipy==1.10.1
torch==1.13.0
torchmetrics==0.10.3
torchsummary==1.5.1
torchvision==0.14.0
tqdm==4.64.1